DAS-2348: Updates HRS to handle SMAP L4 data
Renames _clone_variables -> _clone_dimensions for clarity.

Rewords a number of function comment strings.

_resample_layer is changed a bit, it's now updated to take a fill_value
determined at the variable creation time.

Previously, it called a _prepare_data_plane that was supposed to handle
determining the correct cast and transposition of the input data.  Now we have
turned off the auto_maskandscale for the input and output variables so we are
not relying on any numpy conversions to missing and back.

Now the _resample_layer converts everything (and the fill value to a float64
because pyresample wants that). Then it does the computation using the
fill_value provided rather than the assumed/determined one.

Finally casting and transposing the data back to it's original datatype.

_best_cast is removed because it is unneeded

_resampler_kwargs now only determines the correct rows_per_scan and whether we
should uses maximum_weight_mode (nearest neighbor) if we're integer type data.